+++ src/enum/directoryEnum.js
@@ -1,0 +1,10 @@
+const directoryEnum = {
+  controller: 'controllers',
+  helper: 'Helper',
+  router: 'routers',
+  model: 'models',
+  middleware: 'middlewares',
+  main: 'src'
+};
+
+module.exports = directoryEnum;