+++ src/templates/controller/templateController.js
@@ -1,7 +1,13 @@
+<% if (useTypescript) { %>import { Request, Response } from 'express';  <% } %>
+
 const mainController = {
-  home: function(request, response){
+  home: function(request<% if (useTypescript) { %>: Request <% } %>, response<% if (useTypescript) { %>: Response <% } %>){
     <% if(isApiRest) { %> response.json(`It's alive !`); <% } else { %>response.send(`It's alive !`); <% } %>
   }
 }
 
-module.exports = mainController;
+<% if (useTypescript) { %>
+export { mainController };
+  <% } else { %>
+  module.exports = mainController;
+  <% } %>