+++ src/templates/router/template.router.js
@@ -1,8 +1,16 @@
-const { Router } = require('express');
-const mainController = require('../controllers/mainController');
-
+<% if (useTypescript) { %>
+  import { Router } from 'express';
+  import { mainController } from '../controllers/mainController';
+<% } else { %>
+  const { Router } = require('express');
+  const mainController = require('../controllers/mainController');
+<% } %>
 const mainRouter = Router();
 
 mainRouter.get('/', mainController.home);
 
-module.exports = mainRouter;
+<% if (useTypescript) { %>
+  export { mainRouter };
+<% } else { %>
+  module.exports = mainRouter;
+<% } %>